move vaaRepository to a separate package in order to fix the import cycle
